// src/lib/server/graphqlServer/types/generated-resolvers.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type AccountGroupingSort = {
  active?: InputMaybe<SortDirection>;
  deleted?: InputMaybe<SortDirection>;
  disabled?: InputMaybe<SortDirection>;
  status?: InputMaybe<SortDirection>;
  title?: InputMaybe<SortDirection>;
};
// ...
export type BudgetSort = {
  accountGrouping?: InputMaybe<AccountGroupingSort>;
  active?: InputMaybe<SortDirection>;
  allowUpdate?: InputMaybe<SortDirection>;
  deleted?: InputMaybe<SortDirection>;
  disabled?: InputMaybe<SortDirection>;
  status?: InputMaybe<SortDirection>;
  title?: InputMaybe<SortDirection>;
};
// ...
export type SortDirection =
  | 'asc'
  | 'desc';